Handled validations and cookie based authorization

+++ app/module/routes.py
@@ -2,7 +2,7 @@ from flask import request, jsonify, make_response
 from app import app
 from .models import *
 from .schema import *
-from .const import HttpStatus
+from .const import *
 import bcrypt
 import jwt
 from datetime import datetime, timedelta
@@ -16,23 +16,20 @@ def token_required(f):
     def decorated(*args, **kwargs):
         token = None
         # jwt is passed in the request header
-        print(request.headers)
         if 'auth-token' in request.cookies:
             token = request.cookies['auth-token']
         # return 401 if token is not passed
         if not token:
-            return jsonify({'message': 'Token is missing !!'}), 401
-
+            return response('Token is missing !!',
+                            HttpStatus.UNAUTHORIZED)
         try:
             # decoding the payload to fetch the stored details
             data = jwt.decode(
                 token, app.config['SECRET_KEY'], algorithms=["HS256"])
-            print(data)
             current_user = Users.query.filter_by(userId=data['userId']).first()
         except:
-            return jsonify({
-                'message': 'Token is invalid !!'
-            }), 401
+            return response('Auth Token is Invalid!',
+                            HttpStatus.UNAUTHORIZED)
         # returns the current logged in users contex to the routes
         return f(current_user, *args, **kwargs)
 
@@ -42,31 +39,42 @@ def token_required(f):
 # --------------------- DIVISIONS ENDPOINTS -------------------------
 
 @app.route("/divisions", methods=['GET'])
-def getDivisions():
+@token_required
+def getDivisions(current_user):
     all_divisions = Division.query.all()
     results = divisions_schema.dump(all_divisions)
-    return jsonify(results)
+    return make_response(jsonify(results),
+                         HttpStatus.OK)
 
 
 @app.route("/divisions/<dcode>", methods=['GET'])
-def getDivisionByCode(dcode):
+@token_required
+def getDivisionByCode(current_user, dcode):
     division = Division.query.filter(Division.divisionCode == dcode).first()
-    return division_schema.jsonify(division)
+    if not division:
+        return response("Invalid Division Code Specified!", HttpStatus.INTERNAL_SERVER_ERROR)
+    return make_response(division_schema.jsonify(division),
+                         HttpStatus.OK)
 
 
 @app.route("/divisions/add", methods=['POST'])
-def addDivision():
+@token_required
+def addDivision(current_user):
     divisionName = request.json.get('divisionName')
     technical = request.json.get('technical')
     divisionCode = request.json.get('divisionCode')
-    division = Division(divisionName, technical, divisionCode)
-    db.session.add(division)
-    db.session.commit()
-    return division_schema.jsonify(division)
+    try:
+        division = Division(divisionName, technical, divisionCode)
+        db.session.add(division)
+        db.session.commit()
+        return make_response(division_schema.jsonify(division),
+                             HttpStatus.CREATED)
+    except Exception as e:
+        return response("Invalid Input Payload Supplied", HttpStatus.INTERNAL_SERVER_ERROR)
 
 
 @app.route("/divisions/update/<dcode>", methods=['PUT'])
-def updateDivision(dcode):
+def updateDivision(current_user, dcode):
     division = Division.query.filter(Division.divisionCode == dcode).first()
     divisionName = request.json.get('divisionName')
     divisionCode = request.json.get('divisionCode')
@@ -75,17 +83,23 @@ def updateDivision(dcode):
     division.divisionName = divisionName
     division.divisionCode = divisionCode
     division.technical = technical
-
-    db.session.commit()
-    return division_schema.jsonify(division)
+    try:
+        db.session.commit()
+        return make_response(division_schema.jsonify(division),
+                             HttpStatus.OK)
+    except Exception as e:
+        return response("Invalid Input Payload Supplied", HttpStatus.INTERNAL_SERVER_ERROR)
 
 
 @app.route("/divisions/delete/<dcode>", methods=['DELETE'])
-def deleteDivision(dcode):
+def deleteDivision(current_user, dcode):
     division = Division.query.filter(Division.divisionCode == dcode).first()
+    if not division:
+        return response("Invalid Division Code Specified", HttpStatus.INTERNAL_SERVER_ERROR)
     db.session.delete(division)
     db.session.commit()
-    return division_schema.jsonify(division)
+    return make_response(division_schema.jsonify(division),
+                         HttpStatus.OK)
 
 # ---------------------- USER ENDPOINTS -----------------------
 
@@ -95,31 +109,37 @@ def registerUser():
     username = request.json.get('username')
     password = request.json.get('password')
     emailId = request.json.get('emailId')
-    divisionId = request.json.get('division')
+    divisionId = request.json.get('divisionId')
     firstName = request.json.get('firstName')
     lastName = request.json.get('lastName')
     password = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt(14))
     user = Users(username, password, emailId, divisionId,
                  firstName=firstName, lastName=lastName)
     db.session.add(user)
-    db.session.commit()
-    return "User registered Successfully", 200
+    try:
+        db.session.commit()
+        return response(
+            'User Registered Successfully',
+            HttpStatus.OK
+        )
+    except Exception as e:
+        return response("Invalid Input Payload Supplied", HttpStatus.INTERNAL_SERVER_ERROR)
 
 
 @app.route("/users", methods=['GET'])
 @token_required
-def getUsers():
+def getUsers(current_user):
     users = Users.query.all()
-    return users_schema.jsonify(users)
+    return make_response(users_schema.jsonify(users), HttpStatus.OK)
 
 
 @app.route("/login", methods=['POST'])
 def loginUser():
     username = request.json.get('username')
     user = Users.query.filter(Users.username == username).first()
-    print(user.userId)
     if not user:
-        return "Username Not Found", 403
+        return response("Username Not Found",
+                        HttpStatus.INTERNAL_SERVER_ERROR)
     else:
         password = request.json.get('password')
         if bcrypt.checkpw(password.encode('utf-8'), user.password.encode('utf-8')):
@@ -127,39 +147,45 @@ def loginUser():
                 'userId': user.userId,
                 'exp': datetime.utcnow() + timedelta(minutes=30)
             }, app.config['SECRET_KEY'], algorithm="HS256")
-            print("here")
-            resp = make_response(user_schema.jsonify(user), 201)
+            resp = make_response(user_schema.jsonify(user),
+                                 HttpStatus.OK)
             resp.set_cookie("auth-token", token, httponly=True)
             return resp
         else:
-            return "Invalid Password", 403
+            return response("Invalid Password",
+                            HttpStatus.FORBIDDEN)
 
 
 @app.route("/users/update/<username>", methods=['PUT'])
 @token_required
-def updateUser(username):
+def updateUser(current_user, username):
     user = Users.query.filter(Users.username == username).first()
     if not user:
-        return "User Not Found", 500
+        return response("Username Not Found", HttpStatus.INTERNAL_SERVER_ERROR)
+    elif current_user.userId != user.userId:
+        return response('User not Authorized to update Information', HttpStatus.UNAUTHORIZED)
     else:
         emailId = request.json.get('emailId')
         divisionId = request.json.get('divisionId')
         user.emailId = emailId
         user.divisionId = divisionId
         db.session.commit()
-        return "User Details Updated", 200
+        return response("User Details Updated",
+                        HttpStatus.OK)
 
 
 @app.route("/users/delete/<username>", methods=['DELETE'])
 @token_required
-def deleteUser(username):
+def deleteUser(current_user, username):
     user = Users.query.filter(Users.username == username).first()
     if not user:
-        return "User Not Found", 500
+        return response("Username Not Found",
+                        HttpStatus.INTERNAL_SERVER_ERROR)
     else:
         db.session.delete(user)
         db.session.commit()
-        return "User Deleted Successfully", 200
+        return response("User Deleted Successfully",
+                        HttpStatus.OK)
 
 
 # ------------------ JOBS CONTROLLER---------------------
@@ -167,7 +193,7 @@ def deleteUser(username):
 
 @app.route("/jobs/add", methods=['POST'])
 @token_required
-def addJobs():
+def addJobs(current_user):
     jobTitle = request.json.get('jobTitle')
     postedBy = request.json.get('postedBy')
     isOpen = True
@@ -180,45 +206,59 @@ def addJobs():
     job = Jobs(jobTitle, postedBy, isOpen, jobDescription,
                requirements, salary, lastDateToApply, divisionId)
     db.session.add(job)
-    db.session.commit()
-
-    return "Job Added Successfully", 200
+    try:
+        db.session.commit()
+        return response("Job Added Successfully",
+                        HttpStatus.CREATED)
+    except Exception as e:
+        return response("Invalid Input Payload Supplied",
+                        HttpStatus.INTERNAL_SERVER_ERROR)
 
 
 @app.route("/jobs/delete/<jobId>", methods=['DELETE'])
 @token_required
-def deleteJob(jobId):
+def deleteJob(current_user, jobId):
     job = Jobs.query.filter(Jobs.jobId == jobId).first()
     if not job:
-        return "Job Not Found", 500
+        return response("Job Not Found",
+                        HttpStatus.INTERNAL_SERVER_ERROR)
+    elif current_user.userId != job.postedBy:
+        return response('User is Not Authorized to delete this job',
+                        HttpStatus.UNAUTHORIZED)
     else:
         db.session.delete(job)
         db.session.commit()
-        return "Job Deleted Successfully", 200
+        return response("Job Deleted Successfully",
+                        HttpStatus.OK)
 
 
 @app.route("/jobs", methods=['GET'])
-def getJobs():
+@token_required
+def getJobs(current_user):
     jobs = Jobs.query.all()
-    return jobsSchema.jsonify(jobs)
+    return make_response(jobsSchema.jsonify(jobs), HttpStatus.OK)
 
 
 @app.route("/jobs/update/<jobId>", methods=['PUT'])
 @token_required
-def updateJobs(jobId):
+def updateJobs(current_user, jobId):
     job = Jobs.query.filter(Jobs.jobId == jobId).first()
     if not job:
-        return "Job Not Found", 500
+        return response("Job Not Found", HttpStatus.INTERNAL_SERVER_ERROR)
+    elif current_user.userId != job.postedBy:
+        return response('User is Not Authorized to delete this job',
+                        HttpStatus.UNAUTHORIZED)
     else:
         emailId = request.json.get('emailId')
         divisionId = request.json.get('divisionId')
         job.emailId = emailId
         job.divisionId = divisionId
         db.session.commit()
-        return "Job Details Updated", 200
+        return response("Job Details Updated", HttpStatus.OK)
 
 
 @app.route("/jobs/<jobId>", methods=['GET'])
-def getJobById(jobId):
+@token_required
+def getJobById(jobId, current_user):
     job = Jobs.query.filter(Jobs.jobId == jobId).first()
-    return jobSchema.jsonify(job)
+    return make_response(jobSchema.jsonify(job), HttpStatus.OK)
